CanInitializeMempoolStore
