+++ WalletWasabi.Tests/LiveServerTests.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using WalletWasabi.Backend.Models.Responses;
 using WalletWasabi.Services;
 using WalletWasabi.Stores;
+using WalletWasabi.Stores.Filters;
 using WalletWasabi.Tests.XunitConfiguration;
 using WalletWasabi.TorSocks5;
 using WalletWasabi.WebClients.Wasabi;
