+++ WalletWasabi/Stores/BitcoinStore.cs
@@ -5,6 +5,7 @@ using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 using WalletWasabi.Helpers;
+using WalletWasabi.Stores.Filters;
 
 namespace WalletWasabi.Stores
 {
